Added Validation
Added user validation redirects to the C# of the master page. When run
on the server it will should send you to the correct dash board page.
When run on local host it will fall through and let you go to the page.
Changed teacherDash to use the right master page.