Many things, most untested
Added the delete functionality to the student page. Added update
functionality as well, but commented it out. Made the Defualt master
page use bootstrap and added an error message to Defualt. Made it force
Joe to have teacher permissions. Most of these things I couldn't test
due to lack of data in the DB, so please test them if you have time.